Struct oping::PingItem

pub struct PingItem {
    pub hostname: String,
    pub address: String,
    pub family: AddrFamily,
    pub latency_ms: f64,
    pub dropped: u32,
    pub seq: i32,
    pub recv_ttl: i32,
    pub recv_qos: u8,
}

One ping response from a destination that was added to the Ping context.

Fields

hostname

The hostname as resolved by the library, possibly resolved to a more canonical name.

address

The address as resolved by the library, either IPv4 or IPv6, in textual form.

family

The address family (IPv4 or IPv6) used to ping the destination.

latency_ms

The latency of the response, if any, in milliseconds.

dropped

The dropped-packet count: either 0 or 1.

seq

The sequence number of the ping.

recv_ttl

The TTL on the received response.

recv_qos

The QoS (quality of service) field on the received response.